Windows64 32/64 silliness
On Windows64, long is the same size as int. Use size_t when we
need an actual 64 bit unsigned long.

+++ servers/slapd/config.h
@@ -144,7 +144,7 @@ typedef struct config_args_s {
 		int v_int;
 		unsigned v_uint;
 		long v_long;
-		unsigned long v_ulong;
+		size_t v_ulong;
 		ber_len_t v_ber_t;
 		char *v_string;
 		struct berval v_bv;
